% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Memo.R
\name{Memo}
\alias{Memo}
\title{Memo clean Function}
\usage{
Memo(MEMO, USES)
}
\arguments{
\item{MEMO, }{USES.}
}
\description{
This function splits the MEMO column into MEC1:MEC5
}
\examples{
MEMO()
}
\keyword{MEMO}

