% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParsingFailFix.R
\name{ParsingFailFix}
\alias{ParsingFailFix}
\title{A function to address parsing failures from reading in raw enquiries data}
\usage{
ParsingFailFix(table)
}
\arguments{
\item{table}{input data frame}
}
\description{
This function fixes any parsing fails
}
\examples{
ParsingFailFix()

}
\keyword{Fail,}
\keyword{Parsing}
\keyword{table}

