% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepeatContacts.R
\name{RepeatContact}
\alias{RepeatContact}
\title{A function to calculate repeat contact metrics}
\usage{
RepeatContact(Table)
}
\arguments{
\item{table}{}
}
\description{
This function fixes calculates repeat contact metrics
}
\examples{
RepeatContact()

}
\keyword{contacts,}
\keyword{repeat}
\keyword{table}

