% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SankeyPlot.R
\name{sankeyPlot}
\alias{sankeyPlot}
\title{Create a sankey plot}
\usage{
sankeyPlot(df, name)
}
\arguments{
\item{df}{input dataframe to contruct the Sankey plot}

\item{name}{is the name of the saved HTML sankey to file}
}
\description{
This function creates a sankey plot with an input dataframe
}
\examples{
MEMO()
}
\keyword{Sankey}
\keyword{diagram}
\keyword{plot,}
\keyword{sankey}

