% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamSankey2.R
\name{TeamSankey2}
\alias{TeamSankey2}
\title{Create a sankey plot of memo beginning and end by team}
\usage{
TeamSankey2(df = CallQual3, enquiry, filename)
}
\arguments{
\item{df}{input dataframe of the raw memo data, DEFAULT to CallQual3}

\item{filename}{is the name of the saved HTML of the sankey}

\item{ENQUIRY}{is the Enquiry type to plot in the sankey}
}
\description{
This function creates a sankey plot of an enquiry type beginning and end by team
}
\examples{
TeamSankey2()
}
\keyword{Sankey}
\keyword{diagram}
\keyword{plot,}
\keyword{sankey}

